refactor(bluetooth): Restore class wrappers and split snippets to avoid nesting
