+++ bluetoothle/src/main/java/com/sample/android/bluetoothle/java/BluetoothLeServiceSimple.java
@@ -1,0 +1,59 @@
+/*
+ * Copyright 2026 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.sample.android.bluetoothle.java;
+
+import android.bluetooth.BluetoothAdapter;
+import android.bluetooth.BluetoothDevice;
+import android.bluetooth.BluetoothGatt;
+import android.bluetooth.BluetoothGattCallback;
+import android.bluetooth.BluetoothProfile;
+import android.util.Log;
+
+public class BluetoothLeServiceSimple {
+    private static final String TAG = "BluetoothLeService";
+    private static BluetoothAdapter bluetoothAdapter;
+
+    // [START android_bluetooth_connect_simple_java]
+    public boolean connect(final String address) {
+        if (bluetoothAdapter == null || address == null) {
+            Log.w(TAG, "BluetoothAdapter not initialized or unspecified address.");
+            return false;
+        }
+        try {
+            final BluetoothDevice device = bluetoothAdapter.getRemoteDevice(address);
+        } catch (IllegalArgumentException exception) {
+            Log.w(TAG, "Device not found with provided address.");
+            return false;
+        }
+        // connect to the GATT server on the device
+        return true;
+    }
+    // [END android_bluetooth_connect_simple_java]
+
+    // [START android_bluetooth_callback_simple_java]
+    private final BluetoothGattCallback bluetoothGattCallback = new BluetoothGattCallback() {
+        @Override
+        public void onConnectionStateChange(BluetoothGatt gatt, int status, int newState) {
+            if (newState == BluetoothProfile.STATE_CONNECTED) {
+                // successfully connected to the GATT Server
+            } else if (newState == BluetoothProfile.STATE_DISCONNECTED) {
+                // disconnected from the GATT Server
+            }
+        }
+    };
+    // [END android_bluetooth_callback_simple_java]
+}
